rename the path to esextension again to ensure downards campatibility
