+++ pkg/openSearch/esextension/compositeAgg_test.go
