+++ pkg/openSearch/esextension/scriptedSumAgg.go
