+++ pkg/openSearch/openSearchQuery/compareHandler_test.go
@@ -1,7 +1,7 @@
 package openSearchQuery
 
 import (
-	"github.com/greenbone/opensight-golang-libraries/pkg/openSearch/esextensions"
+	"github.com/greenbone/opensight-golang-libraries/pkg/openSearch/esextension"
 	"testing"
 
 	"github.com/aquasecurity/esquery"
